some work on acgt_gamma.c and Allplots.c now reads h-type and G-type profiles from .profile file and prints lines of different thickness to distinguish them

+++ pynpact/src/extract.c
@@ -44,8 +44,8 @@ char	*argv[];
       exit(1);
    }
 
-   fprintf(stderr,"\nCDS described in line \"%s\" skipping %d+1+%d\n",DESCRIPTION,strlen(DESCRIPTION),SKIP);
-   fprintf(stderr,"\nCDS described in line \"%s\" skipping %d+1+%d\n",DESCRIPTION2,strlen(DESCRIPTION2),SKIP2);
+   fprintf(stderr,"\nCDS described in line \"%s\" skipping %d+1+%d\n", DESCRIPTION, (int)strlen(DESCRIPTION), SKIP);
+   fprintf(stderr,"\nCDS described in line \"%s\" skipping %d+1+%d\n", DESCRIPTION2, (int)strlen(DESCRIPTION2), SKIP2);
 
    while(fgets(longstr,198,input) && !feof(input))
    {
